python/requirements.txt
mysql-connector-python==8.1.0
pip==23.2.1
pytest==7.4.0
